requirements.txt
huggingface_hub>=0.20
python-dotenv>=1.0
tqdm>=4.65
opencv-python-headless>=4.8
numpy>=1.24
torch>=2.1
torchvision>=0.15
torchcodec
transformers>=4.52
matplotlib>=3.7
qwen-vl-utils>=0.0.8
accelerate>=0.26
Pillow>=10.0
